Fix pandas TypeError on type assignment and cast to float
Fixed `TypeError: Invalid value '0.75' for dtype 'int64'` thrown when assigning a float value (`0.75`) to a column originally inferred as `int` or `int64` via `pd.to_numeric(..., errors='coerce').fillna(0)`. We now explicitly cast this column to `float` before making assignments. Fixed tests in `test_klassiekers_sporza.py` and logic in `pages/Klassiekers - Sporza.py` to prevent string to int comparisons and use valid strings for mapping.

+++ pages/Sporza_Giro_Bouwer.py
@@ -132,7 +132,7 @@ def load_all_data():
     naam_col_p = 'Naam' if 'Naam' in df_p.columns else 'Renner'
     naam_col_s = 'Naam' if 'Naam' in df_s.columns else 'Renner'
     df = pd.merge(df_p, df_s, left_on=naam_col_p, right_on=naam_col_s, how='left')
-    df['Prijs'] = pd.to_numeric(df['Prijs'], errors='coerce').fillna(0)
+    df['Prijs'] = pd.to_numeric(df['Prijs'], errors='coerce').fillna(0.0).astype(float)
     df.loc[df['Prijs'] > 1000, 'Prijs'] = df['Prijs'] / 1000000
     df.loc[df['Prijs'] == 0.8, 'Prijs'] = 0.75
     for col in ['GC', 'SPR', 'ITT', 'MTN']:
